follow structure of blogs; send type param to api
