+++ client/src/Layout/Home/Home.jsx
@@ -1,8 +1,8 @@
 import Hero from "./Hero"
 import NewsletterForm from "../Components/NewsletterForm"
 import ProductBanner from "../Components/ProductBanner"
-import About from "./About"
-import UpcomingEvents from "../Components/UpcomingEvents/UpcomingEvents"
+// import About from "./About"
+import Events from "../Events/Events"
 import ContactForm from "../Components/ContactForm"
 import FeaturedBlog from "../Components/FeaturedContent/FeaturedBlog"
 import FeaturedContent from "../Components/FeaturedContent/FeaturedContent"
@@ -11,7 +11,7 @@ export default function Home() {
   return (
     <div>
       <Hero />
-      <UpcomingEvents />
+      <Events />
       <NewsletterForm />
       <ProductBanner />
       {/* <About /> */}
